test(process): route batch argv probe through Windows CI
